More extraction and binary compatibility acceptance
